+++ jest-configs/jest-preprocess.js
@@ -1,0 +1,5 @@
+const babelOptions = {
+  presets: [["@babel/preset-react", { runtime: "automatic" }], "@babel/preset-typescript"],
+};
+
+module.exports = require("babel-jest").createTransformer(babelOptions);