+++ src/components/base/Button/Button.tsx
@@ -1,5 +1,7 @@
 import type { ButtonProps as MuiButtonProps } from "@mui/material";
 import { Button as MuiButton } from "@mui/material";
+import { classNames } from "@src/utils/classNames";
+import test from "./test.json";
 
 // Only include variant, size, and color
 type ButtonBaseProps = Pick<MuiButtonProps, "variant" | "size" | "color" | "className">;
@@ -13,8 +15,15 @@ export interface ButtonProps extends ButtonBaseProps {
 }
 
 export const Button = ({ children, handleClick, ...props }: ButtonProps) => {
+  // test alias - TODO:: remove
+  const onClick = () => {
+    // eslint-disable-next-line no-console
+    console.log(classNames, test.name, test);
+    handleClick?.();
+  };
+
   return (
-    <MuiButton onClick={handleClick} {...props}>
+    <MuiButton onClick={onClick} {...props}>
       {children}
     </MuiButton>
   );